character_generator_1/requirements.txt
Pillow==10.1.0
pip==23.3.2
setuptools==65.5.0
